requirements.txt
PyYAML
Jinja2
pylatexenc
pybtex
